Improve separation of concerns for location manager
* Move fake presenter logic to presenter, isolating fake location
manager with only functions localization (not interface related)
* Nove all fake related stuff to location factory to avoid making cast
outside LocationManager.swift file
